# docs/openapi.yml
openapi: 3.1.0
info:
  title: fast-auth
  description: A simple, extensible authentication library for Axum with JWT and refresh tokens
  license:
    name: Apache-2.0
    identifier: Apache-2.0
  version: 0.1.0
paths:
  /auth/api-keys:
    get:
      tags:
      - auth
      summary: List API keys owned by the current user.
      operationId: api_keys_list
      parameters:
      - name: limit
        in: query
        description: Maximum number of items to return.
        required: true
        schema:
          type: integer
          format: int32
          default: 20
          maximum: 100
          minimum: 1
      - name: offset
        in: query
        description: Number of items to skip from the result set.
        required: true
        schema:
          type: integer
          format: int32
          default: 0
          minimum: 0
      - name: sort_by
        in: query
        description: Column used for sorting.
        required: true
        schema:
          $ref: '#/components/schemas/ApiKeyListSortBy'
      - name: sort_order
        in: query
        description: Sort direction.
        required: true
        schema:
          $ref: '#/components/schemas/ListSortOrder'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListPageResult_ApiKeySummary'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
      security:
      - sessionCookie: []
      - bearerApiKey: []
    post:
      tags:
      - auth
      summary: Create one API key for the current organization.
      operationId: api_key_create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiKeyCreateRequest'
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyCreateResponse'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
      security:
      - sessionCookie: []
      - bearerApiKey: []
...
  /auth/refresh:
    post:
      tags:
      - auth
      summary: Refresh the current session using the refresh-token cookie.
      description: |-
        Rotates the refresh token, issues a new access token, and returns the
        refreshed user payload.
      operationId: refresh
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
        '403':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
  /auth/sign-in:
    post:
      tags:
      - auth
      summary: Sign in an existing user.
      description: |-
        Authenticates user with email and password.
        Sets access and refresh tokens as httpOnly cookies.
        Returns `EmailNotConfirmed` when `require_email_confirmation` is enabled and the user has not
        confirmed their email.
        Calls the `on_sign_in` hook after successful authentication.
      operationId: sign_in
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignInRequest'
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
        '403':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
  /auth/sign-out:
    post:
      tags:
      - auth
      summary: Sign out user by revoking refresh token.
      description: |-
        Revokes the refresh token from database and clears both access and refresh cookies.
        The server automatically sends Set-Cookie headers to clear the cookies from the browser.

        **Note**: Due to the stateless nature of JWT tokens, if the access token was copied before
        sign-out, it will remain valid until it expires (typically 15 minutes). This is standard
        behavior for JWT-based authentication systems.
      operationId: sign_out
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignOutResponse'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
  /auth/sign-up:
    post:
      tags:
      - auth
      summary: Sign up a new user.
      description: |-
        Creates a new user account with email and password.
        When email confirmation is required, it also issues/sends a confirmation email.
        Sets access and refresh tokens as httpOnly cookies unless email confirmation is required.
        Calls the `on_sign_up` hook after successful user creation.
      operationId: sign_up
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignUpRequest'
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
        '409':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
components:
  schemas:
    ApiKeyCreateRequest:
      type: object
      description: Request body for API key creation.
      required:
      - name
      properties:
        name:
          type: string
          description: User-defined API key name.
    ApiKeyCreateResponse:
      type: object
      description: API key creation response with one-time plaintext key.
      required:
      - id
      - name
      - organization_id
      - key
      - created_by_user_id
      - key_prefix
      - created_at
      properties:
        created_at:
          type: string
          format: date-time
          description: Creation timestamp.
        created_by_user_id:
          type: string
          format: uuid
          description: User id that created the key.
        id:
          type: string
          format: uuid
          description: API key identifier.
        key:
          type: string
          description: Plaintext key returned once.
        key_prefix:
          type: string
          description: Stable visible key prefix.
        last_used_at:
          type:
          - string
          - 'null'
          format: date-time
          description: Last successful use timestamp.
        name:
          type: string
          description: User-defined display name.
        organization_id:
          type: string
          format: uuid
          description: Owning organization id.
    ApiKeyListSortBy:
      type: string
      description: Sortable columns for API-key listing.
      enum:
      - created_at
      - name
      - last_used_at
...
    AuthErrorResponse:
      type: object
      description: Standard error response body returned by auth endpoints.
      required:
      - error
      properties:
        error:
          type: string
          description: Human-readable error message.
    AuthResponse:
      type: object
      description: Auth response body. Tokens are set as httpOnly cookies.
      required:
      - user
      - organization
      - auth_role
      properties:
        auth_role:
          type: string
        organization:
          $ref: '#/components/schemas/OrganizationResponse'
        user:
          $ref: '#/components/schemas/UserResponse'
...
    ListPageResult_ApiKeySummary:
      type: object
      description: Shared page payload metadata plus returned items.
      required:
      - items
      - total
      - limit
      - offset
      properties:
        items:
          type: array
          items:
            type: object
            description: API key summary returned by list and delete endpoints.
            required:
            - id
            - name
            - organization_id
            - key_prefix
            - created_by_user_id
            - created_at
            properties:
              created_at:
                type: string
                format: date-time
                description: Creation timestamp.
              created_by_user_id:
                type: string
                format: uuid
                description: User id that created the key.
              id:
                type: string
                format: uuid
                description: API key identifier.
              key_prefix:
                type: string
                description: Stable visible key prefix.
              last_used_at:
                type:
                - string
                - 'null'
                format: date-time
                description: Last successful use timestamp.
              name:
                type: string
                description: User-defined display name.
              organization_id:
                type: string
                format: uuid
                description: Owning organization id.
          description: Returned page items.
        limit:
          type: integer
          format: int32
          description: Applied page size.
          maximum: 100
          minimum: 1
        offset:
          type: integer
          format: int32
          description: Applied page offset.
          minimum: 0
        total:
          type: integer
          format: int64
          description: Total number of visible rows.
    ListSortOrder:
      type: string
      description: Shared sort direction for list endpoints.
      enum:
      - asc
      - desc
...
    OrganizationResponse:
      type: object
      description: Active organization returned in auth responses.
      required:
      - id
      - name
      - role
      properties:
        id:
          type: string
        name:
          type: string
        role:
          $ref: '#/components/schemas/OrganizationRole'
    OrganizationRole:
      type: string
      description: |-
        Organization membership role exposed by auth.

        This role is embedded in [`CurrentUser`], organization membership responses,
        and invitations so backends and HTTP handlers can share one stable
        authorization vocabulary.
      enum:
      - owner
      - admin
      - member
...
    SignInRequest:
      type: object
      description: Request body for sign-in.
      required:
      - email
      - password
      properties:
        email:
          type: string
          description: User email address.
        password:
          type: string
          description: User password.
    SignOutResponse:
      type: object
      description: Response for sign-out.
      required:
      - success
      - message
      properties:
        message:
          type: string
          description: Message instructing client to clear tokens.
        success:
          type: boolean
          description: Sign-out success status.
    SignUpRequest:
      type: object
      description: Request body for sign-up.
      required:
      - email
      - password
      properties:
        email:
          type: string
          description: User email address.
        password:
          type: string
          description: User password (min 8 characters, must contain letter and number).
    UserResponse:
      type: object
      description: User data returned in auth responses.
      required:
      - id
      - email
      properties:
        email:
          type: string
        email_confirmed_at:
          type:
          - string
          - 'null'
        id:
          type: string
  securitySchemes:
    bearerApiKey:
      type: apiKey
      in: header
      name: Authorization
    sessionCookie:
      type: apiKey
      in: cookie
      name: access_token
...